requirements.txt
pytorch-lightning
numpy
scipy
scikit-learn
einops
matplotlib
wandb
omegaconf
hydra-core
pytest
datasets
sphfile
praat-parselmouth
torchinfo
torchmetrics==0.9.0
einops
pytest
matplotlib
pyaudio

# CPC: does not seem to work directly. One must install cython 'pip install cython'
# prior to 'pip install -r requirements.txt'
cython
soundfile
git+https://github.com/facebookresearch/CPC_audio.git